formatting and extraction of Plant into own class
